split WebController into HomeController and EarthquakesController
